extract EntryRecord
