+++ linen-modern/src/main/scala/x7c1/linen/modern/init/unread/AccessorLoader.scala
@@ -5,10 +5,10 @@ import android.database.sqlite.SQLiteDatabase
 import x7c1.linen.glue.res.layout.MainLayout
 import x7c1.linen.modern.accessor.ChannelAccessor.findCurrentChannelId
 import x7c1.linen.modern.accessor.unread.UnreadSourceAccessor
-import x7c1.linen.modern.accessor.{AccountAccessor, AccountIdentifiable, EntryAccessor, EntryAccessorBinder}
+import x7c1.linen.modern.accessor.{AccountAccessor, AccountIdentifiable, EntryAccessor, EntryAccessorBinder, UnreadEntryRow, FooterKind}
 import x7c1.linen.modern.init.unread.AccessorLoader.inspectSourceAccessor
-import x7c1.linen.modern.init.unread.SourceNotLoaded.{AccountNotFound, Abort, ChannelNotFound, ErrorEmpty}
-import x7c1.linen.modern.struct.{UnreadDetail, UnreadOutline, UnreadSource}
+import x7c1.linen.modern.init.unread.SourceNotLoaded.{Abort, AccountNotFound, ChannelNotFound, ErrorEmpty}
+import x7c1.linen.modern.struct.{UnreadDetail, UnreadEntry, UnreadOutline, UnreadSource}
 import x7c1.wheat.macros.logger.Log
 import x7c1.wheat.modern.decorator.Imports._
 import x7c1.wheat.modern.patch.FiniteLoaderFactory
