+++ linen-modern/src/main/scala/x7c1/linen/modern/init/updater/LoadedSource.scala
@@ -1,7 +1,7 @@
 package x7c1.linen.modern.init.updater
 
-import x7c1.linen.modern.accessor.{EntryParts, Updatable}
-import x7c1.linen.modern.accessor.database.SourceRecord
+import x7c1.linen.modern.accessor.Updatable
+import x7c1.linen.modern.accessor.database.{EntryParts, SourceRecord}
 import x7c1.wheat.macros.database.TypedFields
